+++ example/lib/main.dart
@@ -31,14 +31,41 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  List<GanttData<Task>> ganttData = DummyData.data.map((task) {
-    return GanttData<Task>(
+  List<GanttData<Task, Task>> ganttData = DummyData.data.map((task) {
+    return GanttData<Task, Task>(
       dateStart: task.start,
       dateEnd: task.end,
       data: task,
       label: task.name,
+      subData: [
+        GanttSubData<Task>(
+          dateStart: task.start.add(const Duration(days: 1)),
+          dateEnd: task.end,
+          data: task,
+          label: 'Sub ${task.name}',
+        ),
+        GanttSubData<Task>(
+          dateStart: task.start.add(const Duration(days: 1)),
+          dateEnd: task.end,
+          data: task,
+          label: 'Sub ${task.name}',
+        ),
+      ],
     );
   }).toList();
+
+  @override
+  void initState() {
+    ganttData[0].subData.add(
+          GanttSubData<Task>(
+            dateStart: ganttData[0].data.start,
+            dateEnd: ganttData[0].data.end,
+            data: ganttData[0].data,
+            label: ganttData[0].data.name,
+          ),
+        );
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -51,7 +78,7 @@ class _MyHomePageState extends State<MyHomePage> {
               Expanded(
                 child: Padding(
                   padding: const EdgeInsets.all(16.0),
-                  child: GanttChart<Task>(
+                  child: GanttChart<Task, Task>(
                     scrollWhileDrag: true,
                     dayLabelStyle: const TextStyle(
                         fontSize: 14, fontWeight: FontWeight.bold),